--- a/src/Agents.jl
+++ b/src/Agents.jl
@@ -2,6 +2,7 @@ module Agents
 
 using Requires
 using Distributed
+using DistributedArrays
 using DataStructures
 using Graphs
 using DataFrames