This PR evolved from #622 which was attempting to more efficiently use available processors by specifying the `batch_size` kwarg provided by `pmap` as an `ensemblerun!` kwarg. That approach would have forced Agents.jl users to figure out what `batch_size` is optimal, which is not necessarily easy. In this PR I use an alternate approach with the [`DistributedArrays`](https://juliaparallel.org/DistributedArrays.jl/stable/) package that effectively automates setting `batch_size`. In my tests (explained below) I found that the DistributedArray approach outperforms the current `pmap` implementation of `ensemblerun!`. I included both the current `pmap` implementation and my new DistributedArray implementation for now, and will replace `pmap` entirely if you indeed agree that is the correct move.

Using the branch associated with this PR, I ran [this script](https://github.com/mt-digital/JuliaJunk/blob/f28aa721f8c0a8d8785e1849d8851d653f3defff/scripts/schelling_ensemblerun_performance.jl) to test how long different test runs took with the Schelling model, using serial execution, the current parallel execution `pmap` implementation, and a new `darray` version. The Schelling model was run for 100 values of `numagents` and three different `(nreplicates,nsteps)` pairs: `(2, 100)`, `(4,100)`, and `(10,2000)`. Here is how the `models` are constructed that are passed to `ensemblerun!`; `numagents_low = 200` and `numagents_high = 300`:
```julia
basemodels = [Models.schelling(; numagents)[1] 
                          for numagents in numagents_low:numagents_high]
models = repeat(basemodels, nreplicates)
```

The script can be run on the command line, where `-p3` indicates assigning three worker cores for parallelization, like so
```sh
julia -p3 scripts/schelling_ensemblerun_performance.jl
```

Below is the result I obtained on my 8-CPU M1 running Julia 1.7.2 (I obtained similar results on a 2015 MacBook Pro with 4 intel cores on Julia 1.7.1). It shows that with the current implementation of `ensemblerun!`, serial execution outperforms parallel due to excessive data transfer overhead between workers and the coordinating core. However, the `darray` parallel implementation included in this PR is close to serial execution for small numbers of replicates. With ten replicates and 2000 time steps we finally see time savings from the `darray` parallel implementation. This suggests that `darray` is the right approach that will "just work" better at this time. I observed similar speedups in a research model I'm developing. Tests should continue, though, to optimize parallelism in Agents.jl. Happy to clarify or try to simplify anything.
```
ENSEMBLE RUN TEST, n_replicates=2, nsteps=100
serial finished in 0.582458541s
current finished in 6.815945s
darray finished in 1.012393666s
Parallel-to-serial ratio (current): 11.70202601595982
Parallel-to-serial ratio (darray version): 1.738138587961748

ENSEMBLE RUN TEST, n_replicates=4, nsteps=100
serial finished in 0.212536959s
current finished in 11.789640625s
darray finished in 0.234458416s
Parallel-to-serial ratio (current): 55.47101398491356
Parallel-to-serial ratio (darray version): 1.1031418587296151

ENSEMBLE RUN TEST, n_replicates=10, nsteps=2000
serial finished in 9.284566584s
darray finished in 4.186712875s
Parallel-to-serial ratio (darray version): 0.45093250580106986
```