@Datseris 2022-05-28
Thank you very much @mt-digital this is a truly wonderful contribution. I read your PR in detail and I agree that we should follow therefore your approach forwards. Here are two comments:

- Please also include a run with many more models, but much less steps, like 500 replicates and 50 steps 
- For completeness, in your original post in this PR, please edit it and include the above run, but also the results of "current" pmap implementation for the 10 replicates and 2000 steps, because these numbers are not included there.
- In the source code you use the `@DArray` macro. Is it possible to change this version so that it doesn't use macros but normal function and type calls, without any loss of performance? if so, that would be preferred, because macros make source code less readable. If it is not possible to achieve same performance without macros, then it is okay to use the `@DArray` macro.

Besides this comments, I'd say go ahead and remove the old version and add necessary documentation if any, and extra tests if any necessary.