@Datseris 2022-05-28
Ah also I guess this approach needs to be used also for `paramscan` and `ensemlerun` when parallelization is on?