@mt-digital 2022-06-01
Thanks for the feedback. I updated the benchmarks and my initial comment as requested, with some other minor edits.

I do think it's possible to replace the `@DArray` macro; I'll try that in the next few days at most (but hopefully later today).

Regarding `paramscan`, on a first look it appears it won't be possible to keep the progressbar for parallel paramscan and use DistributedArrays. Also, I would probably look to re-use `ensemblerun!` by first initializing an array of models using the `initialize` function argument to `paramscan`, then pass the models array to `ensemblerun!`. I think this would clean up `paramscan.jl` a bit since it could replace the `run_single` helper in both the parallel and serial versions of `paramscan`. 

I think I could restore the progressbar for parallel ensembleruns in a future PR, but for now I would suggest dropping the progressbar for parallel runs. Part of the problem I see is that the [`SharedArrays` of the standard library](https://docs.julialang.org/en/v1/stdlib/SharedArrays/) only allow bits-type shared arrays, which means we can't share `DataFrame` arrays across processors. If `SharedArrays` could be of generic types then we could use a `Distributed` for-loop and track progress there. I believe a `Distributed` for-loop would achieve the same performance as using `DistributedArrays`.