@mt-digital 2022-06-01
Regarding the progressbar for parallel runs, I would be willing to work on that for this PR, but it might take some time.