@Datseris 2022-06-01
i like your suggestion to clean paramscan via utilizing `ensemblerun!` go ahead with that!

I don't fully understand what is the problem with progress bars of parallel runs: my suggestion would be to add a progress bar that counts the overall amount of runs that have been done, not the individual steps within one run (which is what the progress bar in `run!` does)