@Datseris 2022-06-01
oh an also put the benchmarking script you used in a subfolder of the `benchmarks` folder please, for future reference!