@Datseris 2022-06-02
I am not sure how long it should take, because there are a lot of stuff there... But several minutes seems reasonable. Several hours does not! :D