@mt-digital 2022-06-05
Latest push is cleaned up, with everything mostly done. I'll improve documentation and inline comments, then maybe it's close to ready? Let me know if anything is unclear.