@mt-digital 2022-06-14
I'll add to the docs and fix some minor problems by tomorrow. 

The one test on Windows is failing due to an out of memory error on a parallel run, it seems: https://github.com/JuliaDynamics/Agents.jl/runs/6751345338?check_suite_focus=true#step:7:184. Can we bump up the memory or is this fixed? We can reduce memory usage in the test if necessary.