@mt-digital 2022-06-16
OK, I'll reduce the parameter set size and we'll see if that fixes it.

The latest push updates the docs and fixes some coding style issues.