@Datseris 2022-07-04
> @Datseris do we follow any style guide (even roughly) or should code just be clear and readable?

Not yet unfortunately. Soon though I will do a massive PR that restyles all code. For now we can use the SciML one: https://github.com/SciML/SciMLStyle . I agree with almost everything there.