@mt-digital 2022-10-11
Sorry that was just a wip push to save progress, I'll clean up and let you
know when it's done, there's a lot to clean up

On Sat, Jun 4, 2022, 1:38 AM George Datseris ***@***.***>
wrote:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In Project.toml
> <https://github.com/JuliaDynamics/Agents.jl/pull/624#discussion_r889504619>
> :
>
> > @@ -19,18 +19,18 @@ Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
>  ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
>  Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
>  Requires = "ae029012-a4dd-5104-9daa-d747884805df"
> +Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
>
> Neither revise nor StableRNGs need to be in project. If StableRNGs needs
> to be in Project.toml, it must be a test-only dependency.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/JuliaDynamics/Agents.jl/pull/624#pullrequestreview-995786092>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AASQFADWXBA23LE7H4XMCKDVNMIZZANCNFSM5XDLBKZA>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>